calculo do premio para pagamento em maos ou pagseguro informacao de quem autorizou o jogador alteracao do token do pagseguro envio de email de confirmacao de pagamento para autorizacao manual

+++ src/Web/Controllers/BaseController.cs
@@ -61,7 +61,7 @@ namespace BigBallz.Controllers
                 ViewData["LastMatches"] = _matchService.GetLastPlayedMatches();
 
                 var totalprize = _bigBallzService.GetTotalPrize();
-                var prizes = new List<double> {totalprize*0.65, totalprize*0.20, totalprize*0.10};
+                var prizes = new List<decimal> {totalprize*(decimal) 0.65, totalprize*(decimal) 0.20, totalprize*(decimal) 0.10};
                 ViewData["Prizes"] = prizes;
 
                 ViewData["Standings"] = _bigBallzService.GetStandings();
